Testing training, directory pathing, and file renaming code added
Works quite well for adding a training set and testing images against it. JPGs are git ignored, but you can just toss them in the folders. They don't actually need to be renamed, but the folder structure stores the "true label" for the files within it. To Do: add statistics generation.